feat: improve window preview shape
Since not all apps have the same level of window corner rounding, we simply set a transparent background, and display the window as it is instead of interfering with its shape.
In addition, we define the gradient overlay with a mask so that it does not color the transparent corners as well.